metadata_api/requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-Pymong=2.3.0
Werkzeug==1.0.1